+++ agent/event-structs/typed-event-predicates.rkt
@@ -12,15 +12,16 @@
          "message-events.rkt"
          "provider-events.rkt"
          "session-events.rkt"
+         "stream-events.rkt"
          "tool-events.rkt"
          "turn-events.rkt")
 
-(provide
- (all-from-out "base.rkt")
- (all-from-out "hook-events.rkt")
- (all-from-out "iteration-events.rkt")
- (all-from-out "message-events.rkt")
- (all-from-out "provider-events.rkt")
- (all-from-out "session-events.rkt")
- (all-from-out "tool-events.rkt")
- (all-from-out "turn-events.rkt"))
+(provide (all-from-out "base.rkt")
+         (all-from-out "hook-events.rkt")
+         (all-from-out "iteration-events.rkt")
+         (all-from-out "message-events.rkt")
+         (all-from-out "provider-events.rkt")
+         (all-from-out "session-events.rkt")
+         (all-from-out "stream-events.rkt")
+         (all-from-out "tool-events.rkt")
+         (all-from-out "turn-events.rkt"))